+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Switched from OAUTH2 security on FASTAPI to HTTPBearer.
+
 ### Added
 - Add get morphoelectric (me) model tool
+- BlueNaaS simulation tool.
 
 ## [0.1.1] - 26.09.2024